fix(Client-and-Server): Time fields validation fix, for server and client
ABAP date and time fields correspond to either Python string or datetime objects and the configuration which type to use was missing in PyRFC server. The configuration is added for server with default to "strings", the same as for client. The time string plausibility check is now disabled both for client and server, when strings are used, because some ABAP applications may return "240000" time string. Close #336

Closes #336